refactor(calculator): extract data fetching logic to custom hook
- Created usePoolHourlyData hook with encapsulated state
- Removed useEffect #3 from RangeCalculator (data fetching)
- Added cleanup anti-race-condition protection
- Hook now supports configurable daysLookback parameter

Reduces RangeCalculator complexity from 4 to 3 interdependent effects.
Next: Extract hydration and price inversion logic.
